requirements.txt
aniso8601==1.1.0
Flask==0.10.1
Flask-RESTful==0.3.5
Flask-Restless==0.17.0
Flask-SQLAlchemy==2.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
mimerender==0.5.5
passlib==1.6.5
PyJWT==1.4.0
python-dateutil==2.4.2
python-mimeparse==1.5.1
pytz==2015.7
requests==2.9.1
six==1.10.0
SQLAlchemy==1.0.11
virtualenv==14.0.6
Werkzeug==0.11.3
wheel==0.29.0
